+++ src/pcm.h
@@ -38,6 +38,6 @@ void pcm_sample_set_samplerate(uint8_t chan, uint32_t sr);
 void pcm_sample_start(uint8_t chan);
 void pcm_sample_stop(uint8_t chan);
 void pcm_sample_set_loop(uint8_t chan, pcm_sample_loop_t loop);
-void pcm_play(uint8_t chan, SoundFx *sfx, int volume, pcm_sample_loop_t loop);
+void pcm_play(uint8_t chan, SoundFx *sfx, uint8_t volume, pcm_sample_loop_t loop);
 
 void fm_test();
